api/permissions.py
from rest_framework import permissions


class IsManager(permissions.BasePermission):
    def has_permission(self, request, view):
        if not request.user.is_authenticated:
            return False
        return hasattr(request.user, "profile") and request.user.profile.is_manager
