requirements.txt
asgiref==3.8.1
black==24.10.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
filelock==3.16.1
flake8==7.1.1
gunicorn==23.0.0
identify==2.6.1
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.27.1
